重构了Swing Panel和OpenGL Panel的初始化和刷新方式，提高代码可复用性
